feat(users): refactor user service and handlers for better separation of concerns
* extract user creation, deletion, and update logic into dedicated handlers
* implement user query service for fetching user data
* update user service to utilize new handlers and query service
* remove obsolete user service tests
