+++ src/Beagl.WebApp/Program.cs
@@ -4,10 +4,8 @@ using Beagl.Infrastructure;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.AspNetCore.Identity;
 using Beagl.WebApp.Extensions;
-using Beagl.Domain.Services;
 using Beagl.Infrastructure.Entities;
-using Beagl.Infrastructure.Services.Implementations;
-using Beagl.Infrastructure.Services.Interfaces;
+using Beagl.Infrastructure.Extensions;
 
 WebApplicationBuilder builder = WebApplication.CreateBuilder(args);
 
@@ -42,9 +40,9 @@ builder.Services.ConfigureApplicationCookie(options =>
     options.SlidingExpiration = true;
 });
 
-builder.Services.AddScoped<IAuthenticationService, AuthenticationService>();
-builder.Services.AddScoped<IUserService, UserService>();
-builder.Services.AddScoped<IRoleService, RoleService>();
+builder.Services.AddInfrastructureServices();
+builder.Services.AddHandlerServices();
+builder.Services.AddQueryServices();
 
 WebApplication app = builder.Build();
